+++ _templates/package/new/README.md.t
@@ -5,7 +5,6 @@ to: packages/<%= name %>/README.md
 # @feathersjs/<%= name %>
 
 [![CI](https://github.com/feathersjs/feathers/workflows/CI/badge.svg)](https://github.com/feathersjs/feathers/actions?query=workflow%3ACI)
-[![Dependency Status](https://img.shields.io/david/feathersjs/feathers.svg?style=flat-square&path=packages/<%= name %>)](https://david-dm.org/feathersjs/feathers?path=packages/<%= name %>)
 [![Download Status](https://img.shields.io/npm/dm/@feathersjs/<%= name %>.svg?style=flat-square)](https://www.npmjs.com/package/@feathersjs/<%= name %>)
 
 > <%= description %>